+++ app/commands/homebrew_context/daggerheart/change_item_command.rb
@@ -3,6 +3,10 @@
 module HomebrewContext
   module Daggerheart
     class ChangeItemCommand < BaseCommand
+      include Deps[
+        refresh_bonuses: 'commands.bonuses_context.refresh'
+      ]
+
       use_contract do
         config.messages.namespace = :homebrew_item
 
@@ -12,6 +16,11 @@ module HomebrewContext
           optional(:description).maybe(:string, max_size?: 250)
           optional(:info).hash
           optional(:public).filled(:bool)
+          optional(:bonuses).maybe(:array).each(:hash) do
+            required(:id).filled(type_included_in?: [Integer, String])
+            optional(:type).filled(:string)
+            optional(:value)
+          end
         end
       end
 
@@ -23,7 +32,9 @@ module HomebrewContext
       end
 
       def do_persist(input)
-        input[:item].update!(input.except(:item))
+        input[:item].update!(input.except(:item, :bonuses))
+
+        refresh_bonuses.call(bonusable: input[:item], bonuses: input[:bonuses]) if input[:bonuses]
 
         { result: input[:item] }
       end