+++ app/javascript/applications/HomebrewsApp/pages/Content/Daggerheart/Weapons.jsx
@@ -4,7 +4,7 @@ import { createStore } from 'solid-js/store';
 import config from '../../../../CharKeeperApp/data/daggerheart.json';
 
 import { useAppState, useAppLocale, useAppAlert } from '../../../context';
-import { Button, Input, TextArea, Select, createModal, Checkbox } from '../../../components';
+import { Button, Input, TextArea, Select, createModal, Checkbox, ItemBonuses, Bonuses } from '../../../components';
 import { Edit, Trash } from '../../../assets';
 import { fetchDaggerheartBooks } from '../../../requests/fetchDaggerheartBooks';
 import { changeBookContent } from '../../../requests/changeBookContent';
@@ -90,8 +90,10 @@ export const DaggerheartWeapons = () => {
     itemable_id: null,
     description: ''
   });
+  const [itemBonuses, setItemBonuses] = createSignal([]);
   const [selectedIds, setSelectedIds] = createSignal([]);
   const [book, setBook] = createSignal(null);
+  const [bonuses, setBonuses] = createSignal(null);
 
   const [books, setBooks] = createSignal(undefined);
   const [items, setItems] = createSignal(undefined);
@@ -125,6 +127,7 @@ export const DaggerheartWeapons = () => {
   const openCreateItemModal = () => {
     batch(() => {
       setItemForm({ ...itemForm, id: null });
+      setItemBonuses([]);
       openModal();
     });
   }
@@ -147,6 +150,7 @@ export const DaggerheartWeapons = () => {
         itemable_id: null,
         description: item.description.en
       });
+      setItemBonuses(item.bonuses);
       openModal();
     });
   }
@@ -179,19 +183,20 @@ export const DaggerheartWeapons = () => {
   }
 
   const createItem = async (formData) => {
-    const result = await createDaggerheartItem(appState.accessToken, { brewery: formData });
+    const result = await createDaggerheartItem(appState.accessToken, { brewery: formData, bonuses: bonuses() });
 
     if (result.errors_list === undefined) {
       batch(() => {
         setItems([result.item].concat(items()));
         setItemForm({ ...itemForm, id: null });
+        setItemBonuses([]);
         closeModal();
       });
     }
   }
 
   const updateItem = async (formData) => {
-    const result = await changeDaggerheartItem(appState.accessToken, itemForm.id, { brewery: formData, only_head: true });
+    const result = await changeDaggerheartItem(appState.accessToken, itemForm.id, { brewery: formData, bonuses: bonuses(), only_head: true });
 
     if (result.errors_list === undefined) {
       const newItems = items().map((item) => {
@@ -201,13 +206,15 @@ export const DaggerheartWeapons = () => {
           ...formData,
           name: { en: itemForm.name, ru: itemForm.name },
           description: { en: itemForm.description, ru: itemForm.description },
-          features: { en: itemForm.features, ru: itemForm.features }
+          features: { en: itemForm.features, ru: itemForm.features },
+          bonuses: bonuses()
         };
       });
 
       batch(() => {
         setItems(newItems);
         setItemForm({ ...itemForm, id: null });
+        setItemBonuses([]);
         closeModal();
       });
     }
@@ -267,6 +274,7 @@ export const DaggerheartWeapons = () => {
               <td class="p-1" />
               <td class="p-1" />
               <td class="p-1" />
+              <td class="p-1" />
             </tr>
           </thead>
           <tbody>
@@ -294,6 +302,7 @@ export const DaggerheartWeapons = () => {
                     </Show>
                   </td>
                   <td class="minimum-width py-1 text-sm">{item.info.burden}</td>
+                  <td class="minimum-width py-1"><Bonuses bonuses={item.bonuses} /></td>
                   <td class="minimum-width py-1 text-sm">{item.info.features.en}</td>
                   <td class="py-1">{item.description[locale()]}</td>
                   <td>
@@ -402,6 +411,7 @@ export const DaggerheartWeapons = () => {
           value={itemForm.features}
           onChange={(value) => setItemForm({ ...itemForm, features: value })}
         />
+        <ItemBonuses bonuses={itemBonuses()} onBonus={setBonuses} />
         <TextArea
           rows="5"
           containerClassList="mb-2"